The script for the collision detection.
You can simply just copy/paste this into a new Script file within Game Maker Studio 2 to add it. 
The "List" variable is an array with different objects assigned to it that the game will check every single frame for instances of said object and check if there is a collision or not.
The "StartOn" variable is the starting point of the array. For example, if you want to check the first few entries only situationally, this will enable you to do that.